+++ samples/Store/Integration/EventStoreIntegration.fs
@@ -1,14 +1,13 @@
 ﻿[<AutoOpen>]
 module Samples.Store.Integration.EventStoreIntegration
 
-open Equinox.EventStore
+open Equinox.EventStoreDb
 open System
 
-let connectToLocalEventStoreNode log =
-    // NOTE: disable cert validation for this test suite. ABSOLUTELY DO NOT DO THIS FOR ANY CODE THAT WILL EVER HIT A STAGING OR PROD SERVER
-    EventStoreConnector("admin", "changeit", custom = (fun c -> c.DisableServerCertificateValidation()),
-                        reqTimeout=TimeSpan.FromSeconds 3., reqRetries=3, log=Logger.SerilogVerbose log, tags=["I",Guid.NewGuid() |> string]
-    // Connect directly to the locally running EventStore Node without using Gossip-driven discovery
-    ).Establish("Equinox-sample", Discovery.Uri(Uri "tcp://localhost:1113"), ConnectionStrategy.ClusterSingle NodePreference.Master)
+// NOTE: use `docker compose up` to establish the standard 3 node config at ports 1113/2113
+let connectToLocalEventStoreNode (_log : Serilog.ILogger) =
+    let c = EventStoreConnector(reqTimeout=TimeSpan.FromSeconds 3., reqRetries=3, tags=["I",Guid.NewGuid() |> string])
+    // Connect to the locally running EventStore Node using Gossip-driven discovery
+    c.Establish("Equinox-sample", Discovery.ConnectionString "esdb://localhost:2111,localhost:2112,localhost:2113?tls=true&tlsVerifyCert=false", ConnectionStrategy.ClusterSingle EventStore.Client.NodePreference.Leader)
 let defaultBatchSize = 500
 let createContext connection batchSize = EventStoreContext(connection, BatchingPolicy(maxBatchSize = batchSize))