+++ tests/Equinox.EventStore.Integration/Infrastructure.fs
@@ -15,7 +15,11 @@ type FsCheckGenerators =
 #if STORE_POSTGRES || STORE_MSSQL || STORE_MYSQL
 open Equinox.SqlStreamStore
 #else
+#if !STORE_EVENTSTORE_LEGACY
+open Equinox.EventStoreDb
+#else
 open Equinox.EventStore
+#endif
 #endif
 
 [<AutoOpen>]