+++ tools/Equinox.Tool/Program.fs
@@ -226,7 +226,7 @@ let createStoreLog verbose verboseConsole maybeSeqEndpoint =
     let c = LoggerConfiguration().Destructure.FSharpTypes()
     let c = if verbose then c.MinimumLevel.Debug() else c
     let c = c.WriteTo.Sink(Equinox.CosmosStore.Core.Log.InternalMetrics.Stats.LogSink())
-    let c = c.WriteTo.Sink(Equinox.EventStore.Log.InternalMetrics.Stats.LogSink())
+    let c = c.WriteTo.Sink(Equinox.EventStoreDb.Log.InternalMetrics.Stats.LogSink())
     let c = c.WriteTo.Sink(Equinox.SqlStreamStore.Log.InternalMetrics.Stats.LogSink())
     let level =
         match verbose, verboseConsole with
@@ -243,7 +243,7 @@ let dumpStats storeConfig log =
     | Some (Storage.StorageConfig.Cosmos _) ->
         Equinox.CosmosStore.Core.Log.InternalMetrics.dump log
     | Some (Storage.StorageConfig.Es _) ->
-        Equinox.EventStore.Log.InternalMetrics.dump log
+        Equinox.EventStoreDb.Log.InternalMetrics.dump log
     | Some (Storage.StorageConfig.Sql _) ->
         Equinox.SqlStreamStore.Log.InternalMetrics.dump log
     | _ -> ()
@@ -301,7 +301,7 @@ module LoadTest =
             test, a.Duration, a.TestsPerSecond, clients.Length, a.ErrorCutoff, a.ReportingIntervals, reportFilename)
         // Reset the start time based on which the shared global metrics will be computed
         let _ = Equinox.CosmosStore.Core.Log.InternalMetrics.Stats.LogSink.Restart()
-        let _ = Equinox.EventStore.Log.InternalMetrics.Stats.LogSink.Restart()
+        let _ = Equinox.EventStoreDb.Log.InternalMetrics.Stats.LogSink.Restart()
         let _ = Equinox.SqlStreamStore.Log.InternalMetrics.Stats.LogSink.Restart()
         let results = runLoadTest log a.TestsPerSecond (duration.Add(TimeSpan.FromSeconds 5.)) a.ErrorCutoff a.ReportingIntervals clients runSingleTest |> Async.RunSynchronously
 
@@ -315,7 +315,7 @@ let createDomainLog verbose verboseConsole maybeSeqEndpoint =
     let c = LoggerConfiguration().Destructure.FSharpTypes().Enrich.FromLogContext()
     let c = if verbose then c.MinimumLevel.Debug() else c
     let c = c.WriteTo.Sink(Equinox.CosmosStore.Core.Log.InternalMetrics.Stats.LogSink())
-    let c = c.WriteTo.Sink(Equinox.EventStore.Log.InternalMetrics.Stats.LogSink())
+    let c = c.WriteTo.Sink(Equinox.EventStoreDb.Log.InternalMetrics.Stats.LogSink())
     let c = c.WriteTo.Sink(Equinox.SqlStreamStore.Log.InternalMetrics.Stats.LogSink())
     let outputTemplate = "{Timestamp:T} {Level:u1} {Message:l} {Properties}{NewLine}{Exception}"
     let c = c.WriteTo.Console((if verboseConsole then LogEventLevel.Debug else LogEventLevel.Information), outputTemplate, theme = Sinks.SystemConsole.Themes.AnsiConsoleTheme.Code)